requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via azure-functions-durable
aiosignal==1.4.0
    # via aiohttp
attrs==26.1.0
    # via aiohttp
azure-core==1.39.0
    # via azure-storage-blob
azure-functions==2.1.0
    # via
    #   func-auditlog-save (pyproject.toml)
    #   azure-functions-durable
azure-functions-durable==1.5.0
    # via func-auditlog-save (pyproject.toml)
azure-storage-blob==12.28.0
    # via func-auditlog-save (pyproject.toml)
certifi==2026.4.22
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
cryptography==47.0.0
    # via
    #   azure-storage-blob
    #   msal
    #   pyjwt
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
furl==2.1.4
    # via azure-functions-durable
idna==3.13
    # via
    #   requests
    #   yarl
importlib-metadata==8.7.1
    # via opentelemetry-api
isodate==0.7.2
    # via azure-storage-blob
markupsafe==3.0.3
    # via werkzeug
msal==1.36.0
    # via func-auditlog-save (pyproject.toml)
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
opentelemetry-api==1.41.1
    # via
    #   azure-functions-durable
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.41.1
    # via azure-functions-durable
opentelemetry-semantic-conventions==0.62b1
    # via opentelemetry-sdk
orderedmultidict==1.0.2
    # via furl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==3.0
    # via cffi
pyjwt==2.12.1
    # via msal
python-dateutil==2.9.0.post0
    # via azure-functions-durable
requests==2.33.1
    # via
    #   func-auditlog-save (pyproject.toml)
    #   azure-core
    #   azure-functions-durable
    #   msal
six==1.17.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
typing-extensions==4.15.0
    # via
    #   azure-core
    #   azure-storage-blob
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
urllib3==2.6.3
    # via requests
werkzeug==3.1.8
    # via azure-functions
yarl==1.23.0
    # via aiohttp
zipp==3.23.1
    # via importlib-metadata
